feat: swagger for user model

+++ app/Http/Controllers/User/DestroyController.php
@@ -9,11 +9,19 @@ use Symfony\Component\HttpFoundation\Response;
 class DestroyController extends BaseController
 {
     /**
-     * @param User $user
+     * @param int $id
      * @return JsonResponse
      */
-    public function __invoke(User $user): JsonResponse
+    public function __invoke(int $id): JsonResponse
     {
+        $user = User::find($id);
+
+        if (empty($user)) {
+            return response()->json([
+                "message" => "User not found."
+            ], Response::HTTP_NOT_FOUND);
+        }
+
         $this->service->destroy($user);
 
         return response()->json([], Response::HTTP_NO_CONTENT);
